fix: improve ui settings and notion like editor
