+++ web/package.json
@@ -15,6 +15,26 @@
     "@types/react": "^18.2.11",
     "@types/react-dom": "^18.2.4",
     "@types/uuid": "^10.0.0",
+    "@yoopta/accordion": "^4.8.4",
+    "@yoopta/action-menu-list": "^4.8.4",
+    "@yoopta/blockquote": "^4.8.4",
+    "@yoopta/callout": "^4.8.4",
+    "@yoopta/code": "^4.8.4",
+    "@yoopta/divider": "^4.8.4",
+    "@yoopta/editor": "^4.8.4",
+    "@yoopta/embed": "^4.8.4",
+    "@yoopta/exports": "^4.8.4",
+    "@yoopta/file": "^4.8.4",
+    "@yoopta/headings": "^4.8.4",
+    "@yoopta/image": "^4.8.4",
+    "@yoopta/link": "^4.8.4",
+    "@yoopta/link-tool": "^4.8.4",
+    "@yoopta/lists": "^4.8.4",
+    "@yoopta/marks": "^4.8.4",
+    "@yoopta/paragraph": "^4.8.4",
+    "@yoopta/table": "^4.8.4",
+    "@yoopta/toolbar": "^4.8.4",
+    "@yoopta/video": "^4.8.4",
     "axios": "^1.3.6",
     "env-var": "^7.4.1",
     "framer-motion": "^6.5.1",
