+++ web/src/pages/Article.tsx
@@ -13,6 +13,7 @@ import { Anthology, OfflineAnthology } from 'types/anthology';
 import { Article, OfflineArticle } from 'types/article';
 import frenchDate from 'utils/frenchDate';
 import AnthologiesModal from 'components/modals/Anthologies';
+import Editor from 'components/Editor/Editor';
 
 // TODO: redirect if wrong ID
 const ArticlePage = (): JSX.Element => {
@@ -133,12 +134,14 @@ const ArticlePage = (): JSX.Element => {
 								)}
 							</HStack>
 						</VStack>
-						<Text variant="p" whiteSpace="pre-line" textAlign="justify">
-							{user.data.isOffline ? offlineContent : onlineArticle!.content}
-						</Text>
+						<Editor
+							value={user.data.isOffline ? JSON.parse(offlineContent) : JSON.parse(onlineArticle!.content)}
+							readOnly={true}
+						/>
+						<Text variant="p" whiteSpace="pre-line" textAlign="justify"></Text>
 					</VStack>
 					<VStack align="left" spacing="0px" w="100%">
-						// TODO: name author
+						{/* // TODO: name author */}
 						<Text variant="h6">
 							Écrit par {user.data.isOffline ? offlineArticle!.authorId : onlineArticle!.authorId}
 						</Text>
