+++ web/src/theme/components/input.ts
@@ -25,25 +25,25 @@ const input = {
 		primary: {
 			...defaultInput,
 		},
-		'file-primary-orange': {
-			field: {
-				_focus: {
-					border: 'none',
-					outline: 'solid 3px',
-					outlineColor: 'primary.yellow',
-					outlineOffset: '-2px',
-				},
-				_hover: {
-					opacity: '80%',
-					transition: 'ease 0.5s',
-				},
-				borderRadius: 'md',
-				padding: '8px 16px',
-				bg: 'primary.orange',
-				color: 'black',
-				...text.variants.link,
-			},
-		},
+		// 'file-primary-orange': {
+		// 	field: {
+		// 		_focus: {
+		// 			border: 'none',
+		// 			outline: 'solid 3px',
+		// 			outlineColor: 'primary.yellow',
+		// 			outlineOffset: '-2px',
+		// 		},
+		// 		_hover: {
+		// 			opacity: '80%',
+		// 			transition: 'ease 0.5s',
+		// 		},
+		// 		borderRadius: 'md',
+		// 		padding: '8px 16px',
+		// 		bg: 'primary.orange',
+		// 		color: 'black',
+		// 		...text.variants.link,
+		// 	},
+		// },
 	},
 };
