--- web/src/theme/components/switch.ts
@@ -1,14 +1,0 @@
-const cswitch = {
-	variants: {
-		primary: {
-			track: {
-				bg: 'green',
-				_checked: {
-					bg: 'red',
-				},
-			},
-		},
-	},
-};
-
-export default cswitch;
