+++ web/src/theme/index.ts
@@ -3,9 +3,7 @@ import { extendTheme } from '@chakra-ui/react';
 import Text from './components/text';
 import Button from './components/button';
 import Input from './components/input';
-import Textarea from './components/textarea';
 import Select from './components/select';
-import Switch from './components/switch';
 import breakpoints from './foundations/breakpoints';
 // import colors from './foundations/colors';
 import fontWeights from './foundations/fontWeights';
@@ -20,9 +18,7 @@ const overrides = {
 		Text,
 		Button,
 		Input,
-		Textarea,
 		Select,
-		Switch,
 	},
 };
